Move build_ions/els to ParticleSet_builder
