+++ src/Drivers/CMakeLists.txt
@@ -8,7 +8,7 @@ FOREACH(p ${ESTEST})
   TARGET_LINK_LIBRARIES(${p} qmcbase qmcutil ${QMC_UTIL_LIBS} ${MPI_LIBRARY})
 ENDFOREACH(p ${ESTEST})
 
-ADD_LIBRARY(miniwfs ../QMCWaveFunctions/WaveFunction.cpp Mover.cpp)
+ADD_LIBRARY(miniwfs ../QMCWaveFunctions/WaveFunction.cpp)
 
 ADD_EXECUTABLE(miniqmc miniqmc.cpp)
 TARGET_LINK_LIBRARIES(miniqmc miniwfs qmcbase qmcutil ${QMC_UTIL_LIBS} ${MPI_LIBRARY})
